fix bugs

+++ controller/process.go
@@ -145,19 +145,22 @@ func FindLibPathFromPackage(name string) []string {
 }
 
 func (this *Process) Continue() error {
-	// this.UpdatePidList()
+	Continued := make(map[uint32]bool)
 	for _, pid := range this.StoppedPid {
 		// fmt.Printf("Continued pid: %d\n", int(pid))
-		this.MapsUpToDate[pid] = false
-		this.ThreadsUpToDate[pid] = false
-        err := syscall.Kill(int(pid), syscall.SIGCONT)
-        if err != nil {
-            if err == syscall.ESRCH {
-                return fmt.Errorf("No such process -> %d\n", pid)
-            } else {
-                return fmt.Errorf("Process continue error:%v\n", pid)
-            }
-        }
+		if val, ok := Continued[pid]; !ok || !val {
+			this.MapsUpToDate[pid] = false
+			this.ThreadsUpToDate[pid] = false
+			err := syscall.Kill(int(pid), syscall.SIGCONT)
+			if err != nil {
+				if err == syscall.ESRCH {
+					return fmt.Errorf("No such process -> %d\n", pid)
+				} else {
+					return fmt.Errorf("Process continue error:%v\n", pid)
+				}
+			}
+			Continued[pid] = true
+		}
     }
 	this.StoppedPid = []uint32{}
 	this.WorkPid = 0
